feat: Add dataset and data loading from CSV
This commit replaces the hardcoded mock data with data from a real-world predictive maintenance dataset.

- Adds the `ai4i2020.csv` dataset as `src/lib/raw-data.csv`.
- Adds the `papaparse` library for CSV parsing.
- Creates a new `src/lib/data-loader.ts` module to handle reading and parsing the CSV data on the server side.
- Refactors the pages that used the mock data to use the new data loader.
- Moves the client-side logic from the failure analysis page to a new client component to allow for server-side data fetching.
