+++ src/lib/raw-data.csv
@@ -1,0 +1,102 @@
+UDI,Product ID,Type,Air temperature [K],Process temperature [K],Rotational speed [rpm],Torque [Nm],Tool wear [min],Machine failure,TWF,HDF,PWF,OSF,RNF
+1,M14860,M,298.1,308.6,1551,42.8,0,0,0,0,0,0,0
+2,L47181,L,298.2,308.7,1408,46.3,3,0,0,0,0,0,0
+3,L47182,L,298.1,308.5,1498,49.4,5,0,0,0,0,0,0
+4,L47183,L,298.2,308.6,1433,39.5,7,0,0,0,0,0,0
+5,L47184,L,298.2,308.7,1408,40.0,9,0,0,0,0,0,0
+6,M14865,M,298.1,308.6,1425,41.9,11,0,0,0,0,0,0
+7,L47186,L,298.1,308.6,1558,42.4,14,0,0,0,0,0,0
+8,L47187,L,298.1,308.6,1527,40.2,16,0,0,0,0,0,0
+9,M14868,M,298.3,308.7,1667,28.6,18,0,0,0,0,0,0
+10,M14869,M,298.5,309,1741,28.0,21,0,0,0,0,0,0
+11,H29424,H,298.4,308.9,1782,23.9,24,0,0,0,0,0,0
+12,H29425,H,298.6,309.1,1423,44.3,29,0,0,0,0,0,0
+13,M14872,M,298.6,309.1,1339,51.1,34,0,0,0,0,0,0
+14,M14873,M,298.6,309.2,1742,30.0,37,0,0,0,0,0,0
+15,L47194,L,298.6,309.2,2035,19.6,40,0,0,0,0,0,0
+16,L47195,L,298.6,309.2,1542,48.4,42,0,0,0,0,0,0
+17,M14876,M,298.6,309.2,1311,46.6,44,0,0,0,0,0,0
+18,M14877,M,298.7,309.2,1410,45.6,47,0,0,0,0,0,0
+19,H29432,H,298.8,309.2,1306,54.5,50,0,0,0,0,0,0
+20,M14879,M,298.9,309.3,1632,32.5,55,0,0,0,0,0,0
+21,H29434,H,298.9,309.3,1375,42.7,58,0,0,0,0,0,0
+22,L47201,L,298.8,309.3,1450,44.8,63,0,0,0,0,0,0
+23,M14882,M,298.9,309.3,1581,30.7,65,0,0,0,0,0,0
+24,L47203,L,299.0,309.4,1758,25.7,68,0,0,0,0,0,0
+25,M14884,M,299.0,309.4,1561,37.3,70,0,0,0,0,0,0
+26,L47205,L,299.0,309.5,1861,23.3,73,0,0,0,0,0,0
+27,L47206,L,299.1,309.5,1512,39.0,75,0,0,0,0,0,0
+28,H29441,H,299.1,309.4,1811,24.6,77,0,0,0,0,0,0
+29,L47208,L,299.1,309.4,1439,44.2,82,0,0,0,0,0,0
+30,L47209,L,299.0,309.4,1693,30.1,84,0,0,0,0,0,0
+31,M14890,M,299.1,309.5,1339,48.2,86,0,0,0,0,0,0
+32,L47211,L,299.0,309.4,1798,25.5,89,0,0,0,0,0,0
+33,L47212,L,299.0,309.4,1419,48.3,91,0,0,0,0,0,0
+34,L47213,L,298.9,309.3,1665,32.5,93,0,0,0,0,0,0
+35,M14894,M,298.8,309.1,1559,34.7,95,0,0,0,0,0,0
+36,M14895,M,298.8,309.2,1452,48.6,98,0,0,0,0,0,0
+37,M14896,M,298.9,309.2,1581,36.7,101,0,0,0,0,0,0
+38,L47217,L,298.8,309.1,1439,39.2,104,0,0,0,0,0,0
+39,H29452,H,298.9,309.2,1379,50.7,106,0,0,0,0,0,0
+40,L47219,L,298.8,309.1,1350,52.5,111,0,0,0,0,0,0
+41,L47220,L,298.8,309.1,1362,45.4,113,0,0,0,0,0,0
+42,L47221,L,298.8,309.1,1368,50.8,115,0,0,0,0,0,0
+43,M14902,M,298.8,309.1,1368,49.1,117,0,0,0,0,0,0
+44,H29457,H,298.8,309.2,1372,48.5,120,0,0,0,0,0,0
+45,M14904,M,298.8,309.1,1472,47.5,125,0,0,0,0,0,0
+46,L47225,L,298.8,309.1,1489,49.1,128,0,0,0,0,0,0
+47,M14906,M,298.7,309,1843,25.8,130,0,0,0,0,0,0
+48,L47227,L,298.8,309.1,1418,46.3,133,0,0,0,0,0,0
+49,H29462,H,298.8,309.2,1425,53.9,135,0,0,0,0,0,0
+50,M14909,M,298.9,309.2,1412,44.1,140,0,0,0,0,0,0
+51,L47230,L,298.9,309.1,2861,4.6,143,1,0,0,1,0,0
+52,L47231,L,298.9,309.1,1383,54.9,145,0,0,0,0,0,0
+53,H29466,H,298.8,309,1497,43.8,147,0,0,0,0,0,0
+54,L47233,L,298.7,309,1565,35.1,152,0,0,0,0,0,0
+55,L47234,L,298.7,309,1691,30.1,154,0,0,0,0,0,0
+56,L47235,L,298.8,309.1,1512,40.7,156,0,0,0,0,0,0
+57,L47236,L,298.8,309.1,1477,42.4,158,0,0,0,0,0,0
+58,L47237,L,298.8,309.1,1513,40.3,160,0,0,0,0,0,0
+59,M14918,M,298.8,309,1661,29.5,162,0,0,0,0,0,0
+60,L47239,L,298.8,309.1,1378,54.4,165,0,0,0,0,0,0
+61,L47240,L,298.8,309.1,1393,52.6,167,0,0,0,0,0,0
+62,M14921,M,298.8,309,1592,38.2,169,0,0,0,0,0,0
+63,L47242,L,298.8,309,1829,22.9,172,0,0,0,0,0,0
+64,M14923,M,298.9,309,1812,25.9,174,0,0,0,0,0,0
+65,L47244,L,299.0,309.1,1523,38.6,177,0,0,0,0,0,0
+66,L47245,L,299.0,309.1,1394,45.1,179,0,0,0,0,0,0
+67,M14926,M,298.9,308.9,1679,28.7,181,0,0,0,0,0,0
+68,H29481,H,298.8,308.9,1466,44.1,184,0,0,0,0,0,0
+69,L47248,L,298.8,308.9,1527,32.4,189,0,0,0,0,0,0
+70,L47249,L,298.9,309,1410,65.7,191,1,0,0,1,1,0
+71,M14930,M,298.9,309,1924,22.6,193,0,0,0,0,0,0
+72,L47251,L,298.9,309.1,1452,45.5,196,0,0,0,0,0,0
+73,L47252,L,298.9,309.1,1369,44.4,198,0,0,0,0,0,0
+74,L47253,L,299.0,309.1,1592,35.0,200,0,0,0,0,0,0
+75,L47254,L,298.9,309,1601,32.3,202,0,0,0,0,0,0
+76,L47255,L,298.8,308.9,1379,46.7,204,0,0,0,0,0,0
+77,L47256,L,298.8,308.9,1461,47.9,206,0,0,0,0,0,0
+78,L47257,L,298.8,308.9,1455,41.3,208,1,1,0,0,0,0
+79,L47258,L,298.8,308.9,1398,51.5,0,0,0,0,0,0,0
+80,L47259,L,298.8,308.9,1402,37.9,2,0,0,0,0,0,0
+81,H29494,H,298.8,308.8,1445,49.9,4,0,0,0,0,0,0
+82,L47261,L,298.9,308.9,1450,47.5,9,0,0,0,0,0,0
+83,L47262,L,298.9,308.9,1453,40.8,11,0,0,0,0,0,0
+84,L47263,L,299.0,309,1469,40.1,13,0,0,0,0,0,0
+85,L47264,L,298.9,308.9,1422,47.6,15,0,0,0,0,0,0
+86,M14945,M,299.0,309,1804,26.8,17,0,0,0,0,0,0
+87,H29500,H,299.0,309,1436,43.9,20,0,0,0,0,0,0
+88,L47267,L,299.0,309,1544,31.8,25,0,0,0,0,0,0
+89,M14948,M,299.0,309,1419,45.2,27,0,0,0,0,0,0
+90,M14949,M,298.9,308.9,1487,39.5,30,0,0,0,0,0,0
+91,L47270,L,298.9,308.9,1753,25.8,33,0,0,0,0,0,0
+92,L47271,L,298.9,308.9,1830,24.9,35,0,0,0,0,0,0
+93,M14952,M,298.9,308.9,1727,27.1,37,0,0,0,0,0,0
+94,L47273,L,299.1,309,1450,43.0,40,0,0,0,0,0,0
+95,L47274,L,299.0,308.9,1622,31.9,42,0,0,0,0,0,0
+96,M14955,M,299.0,309,1351,52.2,44,0,0,0,0,0,0
+97,M14956,M,299.0,309,1575,35.3,47,0,0,0,0,0,0
+98,M14957,M,298.9,308.9,1750,29.9,50,0,0,0,0,0,0
+99,L47278,L,298.9,308.8,1529,32.7,53,0,0,0,0,0,0
+100,L47279,L,298.8,308.8,1523,44.4,55,0,0,0,0,0,0
+10000,M24859,M,299.0,308.7,1500,40.2,30,0,0,0,0,0,0
